requirements.txt
fastapi
uvicorn[standard]
httpx
pvlib
python-dotenv
pytest
pytest-asyncio
